Fixing issue with NPE
My suspicion is that this works if the classes are compiled in a
specific order. Going out of order should result in an invalid filter
definition being provided. Will investigate to see if this can be
remedied in any way but I personally always include the condition in my
filter definitions for readability and maintainability.